[PRDP-186] Renamed integration test script
